+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/teleops/OneController.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.opmodes;
+package org.firstinspires.ftc.teamcode.opmodes.teleops;
 
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
@@ -11,8 +11,8 @@ import com.seattlesolvers.solverslib.gamepad.ToggleButtonReader;
 import org.firstinspires.ftc.teamcode.Robot;
 import org.firstinspires.ftc.teamcode.util.REVColorSensor;
 
-@TeleOp(name = "Two Controller TeleOp - PS4", group = "Normal Controls")
-public class TwoControllers extends OpMode {
+@TeleOp(name = "One Controller TeleOp - PS4", group = "Normal Controls")
+public class OneController extends OpMode {
 
     private GamepadEx driver1;
     private GamepadEx driver2;
@@ -63,10 +63,10 @@ public class TwoControllers extends OpMode {
 //        driver1.getGamepadButton(GamepadKeys.Button.B).whenPressed(robot.launchColor(Launcher.Color.GREEN));
 //        driver1.getGamepadButton(GamepadKeys.Button.A).whenPressed(robot.launchAll());
 
-        driver2.getGamepadButton(GamepadKeys.Button.SQUARE).whenPressed(robot.launchLeft());
-        driver2.getGamepadButton(GamepadKeys.Button.CROSS).whenPressed(robot.launchMid());
-        driver2.getGamepadButton(GamepadKeys.Button.CIRCLE).whenPressed(robot.launchRight());
-        driver2.getGamepadButton(GamepadKeys.Button.RIGHT_BUMPER).whenPressed(robot.launchAll());
+        driver1.getGamepadButton(GamepadKeys.Button.SQUARE).whenPressed(robot.launchLeft());
+        driver1.getGamepadButton(GamepadKeys.Button.CROSS).whenPressed(robot.launchMid());
+        driver1.getGamepadButton(GamepadKeys.Button.CIRCLE).whenPressed(robot.launchRight());
+        driver1.getGamepadButton(GamepadKeys.Button.RIGHT_BUMPER).whenPressed(robot.launchAll());
 
         // Ascent Lifts - Dpad Up and Dpad Down
 //        driver1.getGamepadButton(GamepadKeys.Button.DPAD_UP).whileHeld(robot.raiseLifts());
